btrfs: forget stray device on failed open
If the physical device of a flakey dm device is tried mounting it fails
to open the device with handle, and leaves behind a stray single device
in the device list.

Remove it if the open fails and if it is a single device. As we don't
register a single device in the device list unless it is mounted.

Signed-off-by: Anand Jain <anand.jain@oracle.com>