btrfs: validate device maj:min during scan
The maj:min of a device can change without altering the device path.
When the device is re-scanned, only the device path change is fixed,
but the changed maj:min remains (bug). This patch fixes it by also
checking for the changed maj:min.

However, please note that we still need to validate the maj:min during
open as in the patch ("btrfs: validate device maj:min during open") because
only the device specified in the mount command gets scanned during mount.

Signed-off-by: Anand Jain <anand.jain@oracle.com>